Add CI support for testing system-modeller (#7)
The workflow runs the full test suite inside the docker-compose environment (thus including MongoDB and Keycloak services). The test results are saved regardless of success. The workflow success/failure is designed to reflect the test success/failure so that other workflows can be chained after it.

As the system-modeller build needs the domain-network artifact from GitHub Packages, it requires a token with package:read rights for the SPYDERISK organisation. This token has been created and stored in the organisation's secrets list.