Improves logging of recommendation algorithm
This is a WIP branch to help understand the behaviour of the recommendations algorithm.
It is not intended to change the behaviour.

I have fixed a bug in the LogicalExpression class on the getListFromOr but I am reasonably confident that the incorrect code was never actually being reached.
I made the LogicalExpression.getListFromAnd into a static method as well, which is correct, but doesn't alter the behaviour.

I experimented with changing the updates in the AttackPathDataset class but ultimately commented out those changes (I am sure they didn't work).

Aside from the logging in the RecommendationsAlgorithm I adjusted the way the applyCSGs method is called to provide a version for the intiial call. The "EMPTY csSet is found" has been changed to a warning as it should not happen (it would mean we were trying to set a CSG which is already activated).

The changes in logback.xml provide a focus on the output from RecommendationsAlgorithm without swamping the logs with other stuff.